separated active & inactive searchTags
also got rid of roboto in a few places
